Added automatic template regression detection.  The template versions are parsed out of the headers of existing generated files.  These template versions are compared against the current templates in use.
If the current template is at a lower version than the template used to generate the previous source, a template regression is detected.  Force must be used to overwrite files with template regressions.

The powershell script has been updated to inform the user in the event of one or more template regressions.

The current regression detection is limited to Controller, View and Resource templates.  The master template version is currently parsed from the header but the regression comparison has not yet been implemented for the master template.